fix: prevent overwriting existing symlink for fd in installation script

+++ setup.sh
@@ -178,7 +178,9 @@ install_fd_find() {
 	if is_ubuntu; then
 		sudo apt install -y fd-find
 		mkdir -p ~/.local/bin
-		ln -s $(which fdfind) ~/.local/bin/fd
+		if [ ! -f ~/.local/bin/fd ]; then
+			ln -s $(which fdfind) ~/.local/bin/fd
+		fi
 	else
 		sudo pacman -Sy --noconfirm fd
 	fi
